Display projected field name in field selection table
