+++ src/components/tables/cells/EditableField.tsx
@@ -1,4 +1,4 @@
-import type { EditableFieldProps } from 'src/components/tables/cells/fieldSelection/types';
+import type { EditableFieldProps } from 'src/components/tables/cells/types';
 
 import { Box, Stack, TableCell, Typography, useTheme } from '@mui/material';
 
@@ -10,8 +10,8 @@ import { useBinding_currentCollection } from 'src/stores/Binding/hooks';
 import { useWorkflowStore } from 'src/stores/Workflow/Store';
 
 export const EditableField = ({
-    buttonStyles,
     field,
+    fieldTextStyles,
     pointer,
     readOnly,
     sticky,
@@ -29,34 +29,36 @@ export const EditableField = ({
             : undefined
     );
 
-    if (field.length === 0 || readOnly) {
-        return (
-            <TableCell sx={sticky ? getStickyTableCell() : undefined}>
-                <Typography>{alternateField ?? field}</Typography>
-            </TableCell>
-        );
-    }
-
     return (
         <TableCell sx={sticky ? getStickyTableCell() : undefined}>
-            <Stack direction="row" spacing={1} style={{ alignItems: 'center' }}>
-                {alternateField ? (
-                    <TextSquare
-                        style={{
-                            color: theme.palette.text.primary,
-                            fontSize: 10,
-                        }}
-                    />
-                ) : (
-                    <Box style={{ height: 15, width: 15 }} />
-                )}
+            {field.length === 0 || readOnly ? (
+                <Typography style={fieldTextStyles}>
+                    {alternateField ?? field}
+                </Typography>
+            ) : (
+                <Stack
+                    direction="row"
+                    spacing={1}
+                    style={{ alignItems: 'center' }}
+                >
+                    {alternateField ? (
+                        <TextSquare
+                            style={{
+                                color: theme.palette.text.primary,
+                                fontSize: 10,
+                            }}
+                        />
+                    ) : (
+                        <Box style={{ height: 15, width: 15 }} />
+                    )}
 
-                <EditProjectionButton
-                    buttonStyles={buttonStyles}
-                    field={alternateField ?? field}
-                    pointer={pointer}
-                />
-            </Stack>
+                    <EditProjectionButton
+                        fieldTextStyles={fieldTextStyles}
+                        field={alternateField ?? field}
+                        pointer={pointer}
+                    />
+                </Stack>
+            )}
         </TableCell>
     );
 };
