+++ src/components/tables/cells/fieldSelection/types.ts
@@ -1,5 +1,4 @@
 import type { ToggleButtonProps, TooltipProps } from '@mui/material';
-import type { CSSProperties } from 'react';
 import type {
     FieldSelectionType,
     TranslatedConstraint,
@@ -8,14 +7,6 @@ import type { FieldSelection } from 'src/stores/Binding/slices/FieldSelection';
 
 export interface ConstraintDetailsProps {
     constraint: TranslatedConstraint;
-}
-
-export interface EditableFieldProps {
-    field: string;
-    pointer: string | undefined;
-    buttonStyles?: CSSProperties;
-    readOnly?: boolean;
-    sticky?: boolean;
 }
 
 export interface FieldActionButtonProps extends ToggleButtonProps {
